requirements.txt
setuptools>=80.9.0
SQLAlchemy==2.0.41
build>=1.2.2.post1
PyMySQL==1.1.1
python-dotenv>=1.1.1